Wp ilib2022

+++ src/main/java/frc/robot/actuators/TalonSRXController.java
@@ -38,8 +38,9 @@ public class TalonSRXController extends WPI_TalonSRX {
     if (!hasEncoder()) {
       return 0;
     }
-    int sensorPosition = this.getSelectedSensorPosition();
-    sensorPosition = (m_subsystemConfig.getBoolean(m_configString + ".encoderInverted") ? -sensorPosition
+    double sensorPosition = this.getSelectedSensorPosition();
+    sensorPosition = (m_subsystemConfig.getBoolean(m_configString + ".encoderInverted")
+        ? -sensorPosition
         : sensorPosition);
     SmartDashboard.putNumber(m_configString + " position", sensorPosition);
     return sensorPosition;
@@ -49,8 +50,9 @@ public class TalonSRXController extends WPI_TalonSRX {
     if (!hasEncoder()) {
       return 0;
     }
-    int sensorVelocity = this.getSelectedSensorVelocity();
-    sensorVelocity = (m_subsystemConfig.getBoolean(m_configString + ".encoderInverted") ? -sensorVelocity
+    double sensorVelocity = this.getSelectedSensorVelocity();
+    sensorVelocity = (m_subsystemConfig.getBoolean(m_configString + ".encoderInverted")
+        ? -sensorVelocity
         : sensorVelocity);
     SmartDashboard.putNumber(m_configString + " velocity", sensorVelocity);
     return sensorVelocity;
